Address PR review (round 2): decompose MRMCPGateway.cpp into Config/Cache/Backend, share spawn helpers, relocate fastmcpp wrapper, switch saveToolsCache to Expected<void>.
Rename dumpToolCacheIfNeeded -> processCmdArgs to keep MCP-arg processing in one place.
