+++ source/MRPch/MRPch.vcxproj
@@ -26,7 +26,6 @@
     <ClInclude Include="MRWasm.h" />
     <ClInclude Include="MRAsyncLaunchType.h" />
     <ClInclude Include="MREigen.h" />
-    <ClInclude Include="MRFastmcpp.h" />
     <ClInclude Include="MRJson.h" />
     <ClInclude Include="MRPch.h" />
     <ClInclude Include="MRTBB.h" />
