+++ source/MRPch/MRPch.vcxproj.filters
@@ -3,7 +3,6 @@
   <ItemGroup>
     <ClInclude Include="MRPch.h" />
     <ClInclude Include="MREigen.h" />
-    <ClInclude Include="MRFastmcpp.h" />
     <ClInclude Include="MRJson.h" />
     <ClInclude Include="MRPch.h" />
     <ClInclude Include="MRTBB.h" />
